+++ NEWS.md
@@ -1,6 +1,13 @@
+# rflow 0.3.4
+
+- fix: fail to init rflow when no data on disk
+- fix: fail when retrieving an rflow element inside cached function
+- add: `element()` function for R6Flow objects
+
+
 # rflow 0.3.3
 
-- `[` and `[[` methods for R6Flow and R6FlowElement
+- `[` method for R6Flow to extract its elements
 
 
 # rflow 0.3.2